test-requirements.txt
coverage
mock
stestr
requests
pytest
ops-scenario>=4.0
